requirements.txt
# FastAPI et serveur
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Base de données
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentification
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.4
joblib==1.3.2

# Utilitaires
python-dotenv==1.0.0
email-validator==2.1.0

# Templates HTML
jinja2==3.1.2

# Graphiques (bonus)
plotly==5.17.0

# Tests (bonus)
pytest==7.4.3
httpx==0.25.2